adding semester to academic curriculum

+++ server/controllers/ModuleController.js
@@ -1,10 +1,30 @@
 const Module = require("../models/ModuleModel");
 const mongoose = require("mongoose");
+const User = require("../models/UserModel");
+const Department = require("../models/DepartmentModel");
+const Curriculum = require("../models/CurriculumModel");
 
 // get all Modules
 const getModules = async (req, res) => {
   const modules = await Module.find({}).sort({ createdAt: -1 });
-  res.status(200).json(modules);
+
+  const module = await Promise.all(
+    modules.map(async (mod) => {
+      const department = await Department.findById(mod.department);
+      const teacher = await User.findById(mod.coordinator);
+      return {
+        ...mod._doc,
+        category: department ? department.departmentName : "Not Categorized",
+        coordinatorTeacher: teacher
+          ? teacher.gender === "Male"
+            ? `Mr. ${teacher.firstName}`
+            : `Mrs. ${teacher.firstName}`
+          : "TBA",
+      };
+    })
+  );
+
+  res.status(200).json(module);
 };
 
 // get a single Module
@@ -25,10 +45,8 @@ const getModule = async (req, res) => {
 
 // create a new Module
 const createModule = async (req, res) => {
-  const {
-    moduleTitle,
-  } = req.body;
-
+  const { moduleTitle, department, coordinator } = req.body;
+  console.log(department, coordinator);
   let emptyFields = [];
 
   if (!moduleTitle) {
@@ -38,14 +56,30 @@ const createModule = async (req, res) => {
   if (emptyFields.length > 0) {
     return res
       .status(400)
-      .json({ error: "Please fill in all the fields", emptyFields });
+      .json({ error: "Please fill in all required the fields", emptyFields });
   }
 
   try {
-    const module = await Module.create({
-      moduleTitle,
-    
-    });
+    const module =
+      !department && !coordinator
+        ? await Module.create({
+            moduleTitle,
+          })
+        : !department
+        ? await Module.create({
+            moduleTitle,
+            coordinator,
+          })
+        : !coordinator
+        ? await Module.create({
+            moduleTitle,
+            department,
+          })
+        : await Module.create({
+            moduleTitle,
+            department,
+            coordinator,
+          });
     res.status(200).json(module);
   } catch (error) {
     res.status(400).json({ error: error.message });
@@ -58,6 +92,15 @@ const deleteModule = async (req, res) => {
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
     return res.status(400).json({ error: "No such Module" });
+  }
+
+  const exist = await Curriculum.find({
+    "grades.subjects.module": id,
+  });
+  if (exist.length !== 0) {
+    return res.status(405).json({
+      error: "Deletion is not allowed once  Subject registered on it!",
+    });
   }
 
   const module = await Module.findOneAndDelete({ _id: id });
